fix(amazonq): delete AwsServerCapabilitiesProvider (#5833)
It is not actually needed; we can read the values from AmazonQLspService
