+++ plugins/amazonq/codewhisperer/jetbrains-community/tst/software/aws/toolkits/jetbrains/services/codewhisperer/CodeWhispererTestBase.kt
@@ -131,7 +131,7 @@ open class CodeWhispererTestBase {
                     get() = CompletableFuture()
 
                 override val initializeResult: Deferred<AwsExtendedInitializeResult>
-                    get() = CompletableDeferred(AwsExtendedInitializeResult())
+                    get() = CompletableDeferred(AwsExtendedInitializeResult(mockk()))
 
                 override val encryptionManager: JwtEncryptionManager
                     get() = TODO("Not yet implemented")
