+++ plugins/amazonq/shared/jetbrains-community/src/software/aws/toolkits/jetbrains/services/amazonq/lsp/model/aws/chat/SendChatPrompt.kt
@@ -4,7 +4,7 @@
 package software.aws.toolkits.jetbrains.services.amazonq.lsp.model.aws.chat
 
 import org.eclipse.lsp4j.TextDocumentIdentifier
-import software.aws.toolkits.jetbrains.services.amazonq.lsp.flareChat.ContextCommand
+import software.aws.toolkits.jetbrains.services.amazonq.lsp.model.aws.ContextCommand
 
 data class ChatParams(
     val tabId: String,
